refactor(admin): extract illuminate layout controller hook (#495 batch 4)
Move the #483 continuous d3-force layout lifecycle out of IlluminateCanvas
into a feature-local controller hook so the component becomes the
Sigma/graphology rendering shell.

- Add app/lib/client/usecase/illuminate/use-illuminate-canvas.ts owning the
  live simulation handle, the rAF tick loop, and the synchronous
  settle/step/pause controls (callback bodies byte-identical to the prior
  in-component driver; stable identities preserved so the once-mounted sigma
  effect's dependency array stays minimal).
- Relocate the pure d3-force module force-layout.ts (+ its test) down from
  components/illuminate/IlluminateCanvas/ to lib/client/usecase/illuminate/
  so the controller hook can import it without the use-case layer importing
  upward from the component layer.
- Trim IlluminateCanvas to drive the hook from the reconcile effect, the drag
  handlers, and the test bridge; refresh the reconcile.ts / force-layout.ts
  header rationale to match the new ownership split.

Behavior-preserving: 485 unit + 48 e2e green (incl. #483/#491/#500/#459
layout-sensitive specs).
